// TPA_WEB_LM/client/src/game/game-interface/action_enum.ts
export enum Action {
    JUMP = "jump",
    MOVING_LEFT = "movingLeft",
    MOVING_RIGHT = "movingRight",
    STOP_RIGHT = "stopRight",
    STOP_LEFT = "stopLeft",
    LOW_KICK = "lowkick",
    FRONT_KICK = "frontkick",
    FRONT_KICK_INVERSER = "frontKickInvsere",
    INVERSE = "INVERSE",
    NORMAL = "NORMAL",
    COLLIDING_LEFT = "collide left",
    COLLIDING_RIGHT = "collide right",
    NOT_COLLIDING = "not collide",
    FIRST_PLAYER= "first player",
    SECOND_PLAYER = "second player",
    LEFT_FRONT_KICK= "left front kick",
    RIGHT_FRONT_KICK = "right front kick"
}